backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.15.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "bcrypt==4.1.2",
    "blinker==1.9.0",
    "brevo-python==1.1.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "flask==2.3.3",
    "flask-cors==4.0.0",
    "flask-jwt-extended==4.5.2",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.1.1",
    "gunicorn==21.2.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "mako==1.3.9",
    "markupsafe==3.0.2",
    "openai==1.65.4",
    "packaging==24.2",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.1",
    "pydantic-core==2.33.0",
    "pyjwt==2.10.1",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.2",
    "requests==2.31.0",
    "sib-api-v3-sdk==7.6.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "stripe==11.6.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.0",
    "typing-inspection==0.4.0",
    "urllib3==2.3.0",
    "uvicorn==0.27.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
    "werkzeug==2.3.7",
]
